* tp_html: typo in tp_html_apply_regex() declarement * tp_jpeg: added "= NULL" to variable declaration * struct file_handle renamed to struct file_handle_struct just to avoid conflicts with other libraries (fcntl.h) and to avoid breaking code referring to "struct file_handle *" * path_resolution.c: specified (file_handle_t *) assignment to prevent some warnings

+++ src/utils.c
@@ -205,7 +205,7 @@ tagsistant_id tagsistant_ID_extract_from_path(const char *path)
 	if (last[0] != NULL)
 		ID = strtol(last[0], NULL, 10);
 
-	dbg(LOG_INFO, "%s has ID %lu", path, ID);
+	dbg(LOG_INFO, "%s has ID %lu", path, (long unsigned int) ID);
 	return ID;
 }
